requirements.txt
 
web3==6.0.0
mnemonic==0.20.0
eth-account==0.10.1
python-dotenv==1.0.0
aiohttp==3.8.1
